separated RoomCardComponent from DashBoard
